Add search bar Organize files in different directories
